backend/requirements.txt
Django==3.2.15
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
importlib-metadata==1.7.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.1
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.3
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.1
requests==2.26.0